Remove content on separate construction and initialization
Separating construction and initialization is generally frowned upon in
C++, and so the chapter isn't likely to be useful. The material that was
there definitely isn't for the situations where separate construction
and initialization is likely to be used. The material that would be
relevant is covered already in other chapters, like the one on re-using
buffers.